requirements.txt
# Core LLM client
openai>=1.12.0,<2.0.0

# HTTP requests
requests>=2.28.0,<3.0.0

# Load .env files
python-dotenv>=1.0.0,<2.0.0

# Colorama to colorize the terminal
colorama>=0.4.6,<1.0